Updated examples, moved code into LiveChat folder
