+++ lib/LiveChat/Models/BaseModel.php
