+++ lib/LiveChat/Models/Goals.php
