+++ lib/LiveChat/REST/Utils.php
